requirements.txt
#External dependencies
argon2-cffi==16.3.0
peewee==3.5.2
pytest==3.6.4